Moving surface mesh related API into a single .cc file
